requirements.txt
anthropic>=0.40.0
edge-tts>=6.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
requests-oauthlib>=1.3.1
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
